pyproject.toml
[project]
name = "metabolic-graph-rag"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=2.9.0",
    "fastapi>=0.128.5",
    "neo4j>=6.1.0",
    "pandas>=3.0.0",
    "prefect>=2.20.25",
    "pydantic>=2.12.5",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "structlog<24",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "codespell>=2.4.1",
]

[tool.codespell]
skip = "uv.lock,notebooks/*.ipynb"
